backend/requirements.txt
fastapi
uvicorn
fastf1
google-generativeai
python-dotenv
pandas
numpy
slowapi
httpx
pydantic
